refactor(api): separate public/admin/run api code
